Create Gym - 104426N
Kaitokid is afraid of fish and has never been on a lake trip because of it. However, his friends want to surprise him with a lake trip, but they need to find a lake with no fish so that Kaitokid can join them.

Help Kaitokid's friends plan a surprise lake trip by determining which lakes are suitable for him to visit.

You are given the number of fish in each lake. For each lake, determine whether it is suitable for Kaitokid to visit or not. If a lake has no fish, it is suitable for Kaitokid to visit. Otherwise, it is not suitable for him to visit.

Input
The first line of the input contains an integer 𝑇 (1≤𝑇≤104), the number of lakes.

For each lake you are given a single integer 𝑥 (0≤𝑥≤100), representing the number of fish in that lake.

Output
For each lake, print YES if it is suitable for kaitokid to visit; otherwise, print NO.

You may print each letter in any case (YES, yes, Yes will all be recognized as positive answer, NO, no and nO will all be recognized as negative answer).