Create Gym - 101102H
Rami and K of his friends are going to watch a movie tonight. There are only one row that contains some available seats, all other rows are taken. Given the description of this row, Rami wants to know if he can find consecutive seats available to sit with his friends.

Input
The first line of input contains a single integer T, the number of test cases.

The first line of each test case consists of two-separated integers C, K (1 ≤ C, K ≤ 105), the number of seats in the row and the number of friends going with Rami, respectively.

The second line contains a string of C digits, where 0 represents an empty seat, and 1 represents taken one.

Output
For each test case, print a single line with yes if Rami can find a place for him and his friends, otherwise print no.