Create AtCoder - abc191_a .cpp
Takahashi and Aoki are playing baseball. Takahashi is the pitcher, and Aoki is the batter. Takahashi can throw an invisible pitch. When he throws it, the ball moves linearly at a constant speedVm/s, and it becomes invisible between the moment T seconds after throwing and the moment S seconds after throwing (inclusive). The ball keeps moving when it is invisible.If the ball is not invisible at the moment the ball is exactly D m away from Takahashi, Aoki can hit the ball. Otherwise, he cannot hit it. Can Aoki hit the ball?