Create AtCoder - abc175_a .cpp
We have weather records at AtCoder Town for some consecutive three days. A string of length 3, S, represents the records - if the i-th character is S, it means it was sunny on the i-th day; if that character is R, it means it was rainy on that day.Find the maximum number of consecutive rainy days in this period.